[RF] Moving code around to accommodate conditions
LOTS of breaking, but less so than previously. A few fixes still needed to deal
with specific edge cases, but almost at a place where conditions are integrated
correctly.
